Update typescript-next-js plugin. Remove old generator
