--- packages/core/src/generate/generateAppFiles/generatorNextJS/generatorNextJS.ts
@@ -1,9 +1,0 @@
-import { generateUseParamsFileNextJS } from "./generateUseParamsFileNextJS";
-import { generateLinkFileNextJS } from "./generateLinkFileNextJS";
-import { generateUseRedirectFileNextJS } from "./generateUseRedirectFileNextJS";
-
-export const generatorNextJS = {
-  generateUseParamsFile: generateUseParamsFileNextJS,
-  generateLinkFile: generateLinkFileNextJS,
-  generateUseRedirectFile: generateUseRedirectFileNextJS,
-};
