--- packages/core/src/generate/generateAppFiles/generatorNextJS/index.ts
@@ -1,1 +1,0 @@
-export * from "./generatorNextJS";
