+++ packages/core/src/plugins/typescript-next-js/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TypescriptNextJSPlugin";
